normaliziation is located in one file
