+++ src/bp/ml/svm/svm.ts
@@ -4,8 +4,7 @@ import numeric from 'numeric'
 
 import BaseSVM from './addon'
 import gridSearch from './grid-search'
-import normalizeDataset from './normalize-dataset'
-import normalizeInput from './normalize-input'
+import { normalizeDataset, normalizeInput } from './normalize'
 import reduce from './reduce-dataset'
 import { SvmConfig, Data, SvmModel, Report } from './typings'
 import { defaultConfig, checkConfig } from './config'
